sahaya-backend/requirement.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
Flask-JWT-Extended==4.5.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
Werkzeug==2.3.7
